+++ addons/gut/gui/GutControl.gd
@@ -50,15 +50,14 @@ func _ready():
 	_config_gui = GutConfigGui.new(_ctrls.settings_vbox)
 
 	_ctrls.test_tree.hide_root = true
-	# Stop tests from kicking off when the runner is "ready" and
-	# prevents it from writing results file that is used by
-	# the panel.
-	_gut_runner.set_cmdln_mode(true)
+	# Stop tests from kicking off when the runner is "ready" and prevents it
+	# from writing results file that is used by the panel.
+	_gut_runner.ran_from_editor = false
 	add_child(_gut_runner)
 
-	# Becuase of the janky _utils psuedo-global script, we cannot
-	# do all this in _ready.  If we do this in _ready, it generates
-	# a bunch of errors.  The errors don't matter, but it looks bad.
+	# Becuase of the janky _utils psuedo-global script, we cannot do all this
+	# in _ready.  If we do this in _ready, it generates a bunch of errors.
+	# The errors don't matter, but it looks bad.
 	call_deferred('_post_ready')
 
 
@@ -297,6 +296,7 @@ func load_config_file(path):
 	_config.options.selected = ''
 	_config.options.inner_class_name = ''
 	_config.options.unit_test_name = ''
+	_config_gui.load_file(path)
 
 
 # ##############################################################################